AT-5944 debug tool functionality on an separate file atat/debug.py
